moved PricingService::arePricesDifferent to ProductPriceCalculation::arePricesDifferent  - removed PricingService
